fix: Handle null actualTypes in matchesParameterTypes
When binding.getParameterTypes() returns null (which can happen
with certain method bindings), we were getting a NullPointerException
when trying to access actualTypes.length.

This fix adds a null check before accessing the array length. When
actualTypes is null, we match only if the query expects no parameters
(i.e., queryParameterTypes is empty).

This fixes the CI failure where the java-provider was crashing during
analysis due to this NPE.

Signed-off-by: tsanders <tsanders@redhat.com>